refactor: rename `/knowledgebase` to `/docs`
