+++ src/features/docs/actions/embed.ts
@@ -1,6 +1,6 @@
-import { embed as embedFromAdapter } from "../../../server/llm/providers/gemini";
-import { EMBEDDING_DIM, EMBEDDING_MODEL } from "../../../server/config";
-import { EmbedRequest, EmbedResponse } from "../../../server/llm/types";
+import { embed as embedFromAdapter } from "@/server/llm/providers/gemini";
+import { EMBEDDING_DIM, EMBEDDING_MODEL } from "@/server/config";
+import { EmbedRequest, EmbedResponse } from "@/server/llm/types";
 
 // Generate embedding
 export async function embed({ text }: EmbedRequest): Promise<EmbedResponse> {
